doc: kyotoproducts.pdf
fix link to kyotoproducts.pdf to point to Internet

There IS an old version of this presentation in the old 1.1.1 release of
Kyotocabinet under the GPL.[1] It comes with the "preferred form of
modification" source .ppt, which opens with LibreOffice. I did not
figure out what font I needed to install to get the Japaneese on the
last slide to show in LibreOffice.

Anyways, at this time, I'd rather not ship the old pdf.

[1]http://fallabs.com/kyotocabinet/pkg/old/kyotocabinet-1.1.1.tar.gz